Fixed: batching lead to faulty results, crashes and men wielding bananas. (#101)
* Fixed: batching lead to faulty results and crashes.

Without this fix, the rotary embedding code had a buffer overrun for the rotation coefficient tables which would lead to incorrect results and sometimes `CUDA error: an illegal memory access was encountered`. The reason is that we were viewing the input as `position * number of heads` rows of head dimension elements each. So when we had processed all rows for a position, we "incremented" the position and kept going, without taking the result modulus the actual number of rows in the batch. So in the best case we would apply the wrong rotary embedding for batch 1, pos 0 forward, worst case we'd try to apply an embedding for position 2049, which we never generated.

The fix is to stick each batch in its own cuda block and do the right number of rows per batch. One could instead have taken the modulus to find the correct positional ordinal (assuming a contiguous x over the batch dimension), but this solution seemed cleaner.

Also added a batch generation validator and confirmed we now get reasonable-looking results.

Refs #50.

* Minor clean-up.

* Fixed: batch test padding not masked out, leading to funny generations for certain models.