Fixed: batch memory corruption due to overrunning `temp_state` buffer.
Without this fix, column_remap_cuda would try to write its remapped output into a temp buffer sized for a batch size of 1. This would work as long as `bsz*seq_len` was less than 2048, then there'd be a problem.

The fix is to use a smaller chunk size that fits in the buffer when the batch is larger. This should not impact performance much since we still work on roughly the same number of elements in parallel apart from some small rounding error when the max_input_len isn't evently divisible by the batch size. And worst case the user can always just increase `max_input_len`.